Move files/directories from passenger role to apache role
The passenger role actually didn't have anything specific to passenger.
Like at all. It was literally all apache stuff, except for the
dependency on the mtpereira.passenger role.

So I got rid of the passenger role and moved all of its stuff to the
apache role, and replace using our passenger role with mtpereira's
in the playbook template