Add non-interactive to certbot command. Bump version to 0.1.3
Small change but fixes a bug where certbot would hang waiting for user
input if you already had a cert for a domain but tried getting a new
cert with both the old and a new domain.

This happens when, for example, production is setup with a temporary url
and then the real url is added later. A new cert needs to be fetched but
without --non-interactive certbot hangs waiting for confirmation that you
want to replace the old cert (with only one domain) with the new cert
with both the old domain and the new domain.